исправление замечаний
