+++ js/form.js
@@ -1,5 +1,5 @@
 import {resetScale } from './scale.js';
-// import {resetEffects } from './photo.js';
+import {resetEffects } from './effects.js';
 
 const MAX_HASHTAG_COUNT = 5;
 const MAX_COMMENT_LENGTH = 140;
@@ -29,7 +29,7 @@ const showForm = () => {
 const hideForm = () => {
   form.reset();
   resetScale();
-  // resetEffects();
+  resetEffects();
   pristine.reset();
   overlay.classList.add('hidden');
   document.body.classList.remove('modal-open');
