+++ js/main.js
@@ -1,2 +1,3 @@
 import'./form.js';
 import './gallery.js';
+
